refactor: mover interfaces de repositório para domain/repository
