+++ src/main/java/com/sensedia/sample/consents/domain/repository/ConsentHistoryRepository.java
@@ -1,4 +1,4 @@
-package com.sensedia.sample.consents.repository;
+package com.sensedia.sample.consents.domain.repository;
 
 import com.sensedia.sample.consents.domain.model.ConsentHistory;
 import org.springframework.data.mongodb.repository.MongoRepository;
