+++ src/main/java/com/sensedia/sample/consents/domain/repository/ConsentRepository.java
@@ -1,4 +1,4 @@
-package com.sensedia.sample.consents.repository;
+package com.sensedia.sample.consents.domain.repository;
 
 import com.sensedia.sample.consents.domain.model.Consent;
 import org.springframework.data.mongodb.repository.MongoRepository;
