+++ src/main/java/com/sensedia/sample/consents/service/impl/ConsentServiceImpl.java
@@ -10,8 +10,8 @@ import com.sensedia.sample.consents.dto.PageResponseDTO;
 import com.sensedia.sample.consents.exception.ConsentNotFoundException;
 import com.sensedia.sample.consents.exception.DuplicateCpfException;
 import com.sensedia.sample.consents.mapper.ConsentMapper;
-import com.sensedia.sample.consents.repository.ConsentHistoryRepository;
-import com.sensedia.sample.consents.repository.ConsentRepository;
+import com.sensedia.sample.consents.domain.repository.ConsentHistoryRepository;
+import com.sensedia.sample.consents.domain.repository.ConsentRepository;
 import com.sensedia.sample.consents.service.ConsentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
