+++ src/test/java/com/sensedia/sample/consents/service/impl/ConsentServiceImplTest.java
@@ -9,8 +9,8 @@ import com.sensedia.sample.consents.dto.ConsentUpdateDTO;
 import com.sensedia.sample.consents.exception.ConsentNotFoundException;
 import com.sensedia.sample.consents.exception.DuplicateCpfException;
 import com.sensedia.sample.consents.mapper.ConsentMapper;
-import com.sensedia.sample.consents.repository.ConsentHistoryRepository;
-import com.sensedia.sample.consents.repository.ConsentRepository;
+import com.sensedia.sample.consents.domain.repository.ConsentHistoryRepository;
+import com.sensedia.sample.consents.domain.repository.ConsentRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
